backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# DXF Processing
ezdxf>=1.4.3

# Geometry Processing
shapely>=2.0.0

# Spatial Search
scipy>=1.11.0
numpy>=1.24.0

# Graph Processing
networkx>=3.2.0

# Data Validation
pydantic>=2.5.0

# AI Integration (Gemini)
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
